\name{embedExtraFonts}
\alias{embedExtraFonts}
\title{Embeds fonts that are listed in the local Fontmap}
\usage{
  embedExtraFonts(file, format, outfile = file,
    options = "")
}
\arguments{
  \item{file}{Name of input file.}

  \item{outfile}{Name of the output file (with fonts
  embedded). (Default is same as input file)}

  \item{format}{File format. (see \code{?embedFonts})}

  \item{options}{Other arguments passed to
  \code{embedFonts}.}
}
\description{
  Embeds fonts that are listed in the local Fontmap
}

