% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafont.r
\docType{package}
\name{extrafont-package}
\alias{extrafont}
\alias{extrafont-package}
\title{extrafont: Tools for Using Fonts}
\description{
Tools to using fonts other than the standard PostScript fonts. This package makes it easy to use system TrueType fonts and with PDF or PostScript output files, and with bitmap output files in Windows. extrafont can also be used with fonts packaged specifically to be used with, such as the fontcm package, which has Computer Modern PostScript fonts with math symbols.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fbertran/extrafont}
  \item Report bugs at \url{https://github.com/fbertran/extrafont/issues}
}

}
\author{
This package has been written by Winston Chang (github.com/wch/extrafont)

Maintainer: Frederic Bertrand <frederic.bertrand@lecnam.net>
}
\keyword{internal}
