\name{font_addpackage}
\alias{font_addpackage}
\title{Add font from an installed package to extrafont's registry}
\usage{
  font_addpackage(pkg = NULL)
}
\arguments{
  \item{pkg}{The name of the font package, e.g.,
  \code{"fontcm"}.}
}
\description{
  Add font from an installed package to extrafont's
  registry
}

