% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonttable.r
\name{fonts}
\alias{fonts}
\title{Show the fonts that are registered in the font table
(and available for embedding)}
\usage{
fonts()
}
\description{
Show the fonts that are registered in the font table
(and available for embedding)
}
