% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.r
\name{loadfonts}
\alias{loadfonts}
\title{Reads the fonttable database and registers those fonts with R}
\usage{
loadfonts(device = c("all", "pdf", "postscript", "win"), quiet = FALSE)
}
\arguments{
\item{device}{The output device. If \code{"all"}, then it will load \code{"pdf"}, \code{"postscript"}, and \code{"win"} (if on Windows).}

\item{quiet}{If \code{FALSE}, print a status message as each font
is registered. If \code{TRUE}, don't print.}
}
\value{

}
\description{
This registers fonts so that they can be used with the pdf,
postscript, or Windows bitmap output device. It must be run
once in each R session.
}
\seealso{
\code{\link{embed_fonts}},
#ifdef windows
  \code{\link[grDevice]{windowsFont}}, \code{\link[grDevice]{windowsFonts}},
#endif
  \code{\link{postscriptFonts}},
  \code{\link{pdfFonts}}, \code{\link{Type1Font}}.
}
