\docType{package}
\name{extrafont}
\alias{extrafont}
\alias{extrafont-package}
\alias{package-extrafont}
\title{extrafont package}
\description{
This package is used for using and embedding fonts other
than the basic Postscript fonts.
}
\details{
For use instructions, see
\url{https://github.com/wch/extrafont}.
}

