\name{font_addpackage}
\alias{font_addpackage}
\title{Add font from an installed package to fonts database}
\usage{
font_addpackage(pkg = NULL)
}
\arguments{
  \item{pkg}{The name of the font package, e.g.,
  \code{"fontcm"}.}
}
\description{
This is usually not called directly, but via
\code{\link{font_install}}.
}
\seealso{
\code{\link{font_install}}
}

