% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.r
\name{font_import}
\alias{font_import}
\title{Import system fonts}
\usage{
font_import(paths = NULL, recursive = TRUE, prompt = NULL, pattern = NULL)
}
\arguments{
\item{paths}{A vector of directories to search in. (Default is to auto-detect based on OS)}

\item{recursive}{Search recursively in directories? (Default TRUE)}

\item{prompt}{Show confirmation prompt? (Default TRUE for interactive sessions)}

\item{pattern}{A regular expression that the filenames must match.}
}
\description{
Presently only supports TrueType fonts.
}
\examples{
font_import()

}
