\docType{package}
\name{fonts}
\alias{fonts}
\alias{fonts-package}
\alias{package-fonts}
\title{fonts package}
\usage{
  fonts()
}
\description{
  This package is used for using and embedding fonts other
  than the basic Postscript fonts.

  Show the fonts that are registered in the font table (and
  available for embedding)
}
\details{
  For use instructions, see
  \url{https://github.com/wch/fonts}.
}

