\name{ttf_import}
\alias{ttf_import}
\title{Imports all TrueType fonts in a directory and all subdirectories}
\usage{
  ttf_import(paths = NULL, recursive = TRUE)
}
\arguments{
  \item{paths}{A vector of directories to search in.
  (Default is to auto-detect based on OS)}

  \item{recursive}{Search recursively in directories?
  (Default TRUE)}
}
\description{
  Imports all TrueType fonts in a directory and all
  subdirectories
}

